Add Input module
This represents the source of the data to be plotted.
