+++ hvega/src/Graphics/Vega/VegaLite.hs
@@ -606,6 +606,7 @@ import Prelude ()
 import qualified Graphics.Vega.VegaLite.Configuration as VL
 import qualified Graphics.Vega.VegaLite.Core as VL
 import qualified Graphics.Vega.VegaLite.Data as VL
+import qualified Graphics.Vega.VegaLite.Input as VL
 import qualified Graphics.Vega.VegaLite.Output as VL
 import qualified Graphics.Vega.VegaLite.Selection as VL
 import qualified Graphics.Vega.VegaLite.Specification as VL
